Simplify Rust and Toml code style settings
Previously, each language provided two extensions: `codeStyleSettingsProvider` and `langCodeStyleSettingsProvider`. And it worked for some reason.
But it doesn't make sense since `LanguageCodeStyleSettingsProvider` is inherent of `CodeStyleSettingsProvider` and it can provide all necessary things itself.

So these changes merge both provides into single one. Also, `createSettingsPage` is replaced with `createConfigurable` because `createSettingsPage` is deprecated
