+++ intellij-toml/src/main/kotlin/org/toml/ide/formatter/settings/TomlLanguageCodeStyleSettingsProvider.kt
@@ -5,14 +5,33 @@
 
 package org.toml.ide.formatter.settings
 
+import com.intellij.application.options.CodeStyleAbstractConfigurable
+import com.intellij.application.options.CodeStyleAbstractPanel
 import com.intellij.application.options.IndentOptionsEditor
 import com.intellij.application.options.SmartIndentOptionsEditor
 import com.intellij.lang.Language
+import com.intellij.psi.codeStyle.CodeStyleConfigurable
+import com.intellij.psi.codeStyle.CodeStyleSettings
+import com.intellij.psi.codeStyle.CustomCodeStyleSettings
 import com.intellij.psi.codeStyle.LanguageCodeStyleSettingsProvider
 import org.toml.lang.TomlLanguage
 
 class TomlLanguageCodeStyleSettingsProvider : LanguageCodeStyleSettingsProvider() {
     override fun getLanguage(): Language = TomlLanguage
+
+    override fun createCustomSettings(settings: CodeStyleSettings): CustomCodeStyleSettings =
+        TomlCodeStyleSettings(settings)
+
+    override fun createConfigurable(
+        baseSettings: CodeStyleSettings,
+        modelSettings: CodeStyleSettings
+    ): CodeStyleConfigurable {
+        return object : CodeStyleAbstractConfigurable(baseSettings, modelSettings, configurableDisplayName) {
+            override fun createPanel(settings: CodeStyleSettings): CodeStyleAbstractPanel =
+                TomlCodeStyleMainPanel(currentSettings, settings)
+        }
+    }
+
     override fun getCodeSample(settingsType: SettingsType): String =
         when (settingsType) {
             SettingsType.INDENT_SETTINGS -> INDENT_SAMPLE
